Trigonometry shorthands introduced by Red
This PR contains the abbreviations for the trigonometry functions
as introduced by Red (updated from the post, e.g. it was initially
ARCTAN for the arctangent, but changed to ATAN)

http://www.red-lang.org/2014/08/043-floating-point-support.htm

Raising a question about whether to do this or not is that there
is contention between the color code for TAN and the abbreviated
form of TANGENT.  It's not clear which would be more deserving of
a "global" name, or if either colors or math functions should be put
under some more specific context.

In any case, this raises the question...and also updates some of
the help strings (as well as correcting mistakes from Red, which
at time of writing lists the ARC* functions as taking an angle)