Rebol-based REPL (abandoned to advance GUI debugger)
This commit contains the two userspace files from an abandoned attempt
to write the Rebol-REPL-in-REPL.  The main problem with such a choice
had to do with its impacts on the debugger, when Rebol does not offer
something like V8's "isolates" (where even though V8 is single threaded
it can still maintain independence of multiple executing interpreters).

Writing a viable debugger if the REPL was in userspace is not really
possible at the moment, as the stack of the REPL and the stack of
the executing program would be too intertwined.  This would manifest
in ways that are kludgey to work around (such as filtering stack
backtraces to hide REPL implementation frames), as well as ways that
begin to seem impossible (such as handling step-over and step-out
instructions without ever accidentally crossing into the REPL.)

But also, most people's ideal of a debugger isn't in a REPL anyway.
It would be better to have GUI windows displaying a list of stack
frames and visible breakpoint markers on source.  For Ren-C, the
shortest path would be to write that debugger as part of Ren Garden as
C++/Qt.

In the interests of not holding back the Ren Garden debugging experience
for the sake of having Rebol's REPL written in Rebol, the internal
support code for implementing this REPL was abandoned.  However, this
branch preserves the usermode REPL itself in case it becomes interesting
to look at when/if Rebol gets "isolates" like V8's.

One of the more interesting aspects of the REPL was that it achieved
multi-line input by reading lines, and not DOing so long as there
is an "unmatched delimiter" error when it tried LOADing the
accumulated input thus far.  A C-based REPL built on Ren-C instead of
the R3-Alpha host kit could use the same tactic.