Allow enfixing of functions whose first argument is variadic
Previously it was illegal to enfixing a function whose first argument
is a VARARGS!, as you clearly can't write:

    1 + 2 4 + 5 my-enfixed-op

...with the expectation that it can take backwards twice and get 9 and
3.  The result of 1 + 2 is discarded during left to right evaluation
before the enfixed op gets a chance to see it.  Trying this would cause
an assert.

Rather than turn this assertion to an error, this commit implements a
potentially more interesting behavior.  It allow the left hand side of
an enfix to be a source of "0 or 1" arguments for a VARARGS!.  This is
identical to what is possible with an `<end>`-able argument, however it
means you don't have to rewrite a variadic function if this behavior
suits you.

Note: This is a bit shady (in cases besides an <end> on the left being
a varargs that reports TAIL? as TRUE).  That's because most variadics
expect their evaluation to happen when they TAKE a VARARGS!, and not
beforehand.  But you can't defer the evaluation of a left-hand
expression, because it's usually too late. Even if it isn't technically
too late for some reason (e.g. it's #tight, or quoted) there's still a
bit of an oddity, because variadics on the right have the option to
*not* do a TAKE and leave the value for consumption by the next
operation.  That doesn't apply when the variadic is being "faked in"
from the left.

But despite the lack of "purity", one might argue it's better to do
something vs. just give an error.  Especially since people are unlikely
to enfix a variadic on accident, and may be fine with these rules.