No transparent BOM handling, kill off TO-STRING/ASTRAL
The TO-STRING/ASTRAL code was a stopgap measure to permit a user hook
when unsupported codepoints were hit.  It won't be necessary with
UTF-8 Everywhere.

There was also a wrapper for the UTF16BE, UTF16LE, and UTF8 decoding
which should only ever be called now with UTF8.  This just goes ahead
and gets rid of that, and switches the tests based on UTF16BE and
UTF16LE to use the respective codecs of LOAD (instead of expecting
auto-detection in READ).

Consequently, this commit had to deal with some issues related to tests
that were expecting BOM handling.  It went with the decision that BOM
handling should not be transparent in the core conversions, but should
be the responsibility of codecs...and if the BOM makes it into a string
it should mold visibly as being escaped:

https://github.com/rebol/rebol-issues/issues/2280

Also switches Append_Unencoded_XXX() to only take byte sized targets,
as it is mostly used with mold.  An exceptional case of error forming
is switched away from building strings directly and trying to
Append_Unencoded to them, and goes via the mold buffer instead.