Soften notion that all UTF-8 data must be REBYTE*
Working with UTF-8 bytes requires a fixed notion of sign, which is
not given by C's `char` (it is signed on some platforms and unsigned
on others).  So the idea of using the type difference between `REBYTE*`
and `char*` seemed like a good idea, especially since the C strlen()
function doesn't reflect Rebol's notion of the "length" of a string
when expressed as UTF-8.

However, at the edges of the API where UTF-8 bytes are not manipulated
explicitly, it is burdensome to impose a separate datatype.  In fact,
the datatype "char" can be taken in the inverse meaning--that a pointer
to it explicitly means the data is in UTF-8 format.

These two notions compete somewhat, but there is a likely compromise
at the edges of the API.  Also, the size_t type can be used to convey
a difference between a size in bytes and a length in characters.

(This actually reverses some recent "correctness" on filenames that
were UTF8 but not REBYTE*.  Hammering out these definitions will be an
ongoing process.)