New REBCHR macro for safe UTF-8 enumeration
This retakes the name of the recently-eliminated REBCHR in order to make a
new character enumeration class that can be used to visit characters in a
UTF-8 string.  The C++ build is able to treat it as a class that has
increment and decrement disabled, forcing the usage of helpers...while in
the C build, it is just a plain data type.  Technique described here:

http://blog.hostilefork.com/kinda-smart-pointers-in-c/

The goal is to convert all non-codepoint applications of REBUNI to REBCHR(),
allowing it to continue processing 16-bit characters in the meantime...then
flip the switch to use variable-width characters.