Write string reversal in variable size encoding form
This implements partial string reversal by using a temporary buffer,
and then calling the same code as CHANGE/PART in order to overwrite the
portion of the string (index to end) that was requested to be reversed.

This also temporarily disables sorting the individual characters in strings
(an uncommon operation) rather than try and write an in-place UTF-8-safe sort;
which can't in the general case use C's qsort().  (Only if the string is
all ASCII...)

Shuffling is also an uncommon operation, punted on for the moment.