Deline strings at scanner time, remove Enline/Deline functions
Scanning strings was being done in two steps...one to walk through the UTF-8
data and read it into a string buffer, and a second sweep to walk through
the bytes and correct any CRs to LF or CR LF to LF.  It actually is the case
that the UTF-8 reading routine itself was able to do this "delining", it
just hadn't exposed a parameter for it.

Correcting this is not only more efficient, it reduced the number of places
ENLINE and DELINE were needed to a single native each, allowing the code and
interface to be minimized to make it easier for variable-width character
encoding.