Split_Lines rewrite, disable "raw print"
While SPLIT is usermode, READ/LINES uses an internal implementation, which
needs to be able to accomodate variable width characters in UTF-8.  This uses
a method which should be more easily adaptable to that, which is able to
count characters in each string being split out (though currently it's just
doing byte sizes).

In preparation for speaking to the print interface purely in Rebol types,
this removes the "raw" encoding option.  The datatype (BINARY! vs. STRING!)
would be used for this distinction.

Also simplifies the interface for NEXT_CHR and BACK_CHR such that it doesn't
need separate const versions (const_NEXT_CHR and const_BACK_CHR).  Instead it
uses a mutability cast to return a pointer that shouldn't be written to (the
C++ build enforces it by way of overloaded methods)