Change ANY-STRING! to use byte-sized series, but still UCS2
This is a gentle shift toward the UTF-8 everywhere goal, to start
introducing the necessary terminology and features of distinguishing
"length" in characters from "size" in bytes.  However, it maintains a
`size = length * 2` invariant for the moment...keeping the characters
encoded as UCS2 data.  It just prepares the system at various edge
points for the change.

The REBSER node now speaks of its length in series units that are
currently in use as `used`, not `len`.  A position in bytes is an
`offset`...compared to an `index` which is measured in series units.

To help with the checking, it introduces a temporary information flag
SERIES_INFO_UCS2_STRING.