Use mold buffer for FILE/LOCAL transforms
R3-Alpha had a fair amount of "string surgery" by which it would first
copy a file path, then walk over it with GET_ANY_CHAR and SET_ANY_CHAR.
It also would pre-build-in two additional characters worth of "extra
space" in file conversions to sneakily allow for wildcards to be
added in to them, since listing directories requires appending a "*"
in Windows to the directory path.

Because such indexed manipulations are not generically easy or fast to
do on UTF-8 strings, this switches the implementation to use the mold
buffer.  It also doesn't do directory transformations on FILE!s before
calling the device layer (e.g. adding a `*` to them), especially since
POSIX doesn't even *want* the `*` (it had to remove it).  Instead,
the File-to-Local transformation has options that the relevant code
for each OS can use, such as asking for the wildcard specifier or for
the directory to be given without a trailing slash.