Make %t-binary.c separate from %t-string.c
This takes the dispatching code for the BINARY! type and removes it
from being a special case of the STRING! dispatcher.  At this point,
the types are too different to meaningfully process in the same
code (STRING!s however are becoming more like WORD!s, so those might
unify instead.

Additionally, this changes the way series "common" actions are
dispatched so that the REMOVE action can be handled specially by
strings.

Has enough changes in it that the interpreter can make it through boot.