Update Find_XXX functions, other changes to get tests to run
This is a batch of changes which is sufficient to get the tests to run
all the way through without asserting or crashing.

The most significant part of this is the changes to the various
functions used to FIND types in binaries and series.  The code was
shared between the FIND action and PARSE, but a lot of work was still
repeated.  In addition to fixing them to be able to scan through
variable-size characters, this reduces the redundant code paths a bit.

It also works on formalizing the semantics of what it means to look
for strings in binaries.  Since the UTF-8 everywhere commitment means
strings are represented in binary as UTF-8, that's what is used...and
the allowance is made to find UTF-8 patterns even in binary data that
is not completely UTF-8.