Const correctness for VAL_UTF8_AT and VAL_BYTES_AT
Several byte-oriented routines were willing to accept either BINARY!
values, or STRING! values which would be considered the same as their
UTF-8 binary equivalents.  This saved people from having to convert
arguments to BINARY! to use them as binary parameters.

(This may be somewhat questionable, given that AS BINARY! is a fairly
cheap operation for ANY-STRING!.)

Because these routines were willing to give access to byte data inside
of WORD!s, and read-only strings, this access should be const.  But
the return type was a mutable byte pointer.  None of the clients
needed to write (nor should they have, as it would have been buggy),
so this changes the return type to const.

The only questionable case is the FFI, which gives out pointers to
STRING! and BINARY! data, but can't enforce them as read-only.