Replace /LAST /REVERSE on FIND and SELECT with -LAST, -REVERSE
The /REVERSE refinement of FIND and SELECT would theoretically run
with a /SKIP of -1, while the /LAST refinement would imply a /SKIP of
-1 along with starting the search from the tail of the series.

This meant both refinements were incompatible with /SKIP, and also the
use of /REVERSE with /LAST is redundant.  Additionally, the multiple
refinement behaviors were masking bugs and missing features with /SKIP
(which seemed largely unimplemented).

This commit addresses the issues by making FIND-REVERSE and FIND-LAST
specializations and adaptations of a basic FIND function, which offers
just the /SKIP refinement.  This puts more pressure on fixing bugs in
the base function, with less C code to make mistakes interpreting the
refinements across types.