Get rid of Debug_Fmt(), replace with rebPrint() or printf
One of the premises of Rebol was not to depend on standard library
functions like printf.  However, there was a routine very much like
printf written in the C code, which had some adaptations to let it
print out Rebol values/series.

Despite its name, it was included in release builds as well.  It
provided the string output and formatting for tracing routines that
users might be interested in using, and which should be going through
whatever I/O methods their console uses.

Ren-C compromises on the use of printf() by allowing it in debug
builds.  This makes it possible to instrument Rebol during early boot
moments (or otherwise malfunctioning systems) without going through
any PORT! or print machinery.  However, release builds do not link in
printf(), which avoids carrying around code that isn't exposed to
userspace (and shouldn't be).

The debug-build-only usages of Debug_Fmt have been changed to usages
of printf() and the polymorphic PROBE().  This takes the next step
and makes all the release-build usages switch over to using rebPrint().