Restore FFI functionality
The FFI extension was not being built since Dec 8, 2008...due to build
system changes that stopped it from compiling.  Since that time,
several things changed...including REB_STRUCT no longer being one of
the 64 built-in types.  This led to major code changes that were
roughed in, but never compiled or tested.

This commit involves getting the extension into a compiling state,
as well as making sure all the tests run.  It also moves those tests
into a subdirectory of the extension (as well as some tests that are
only relevant if other extensions are built in, to their corresponding
extension directories.)

One feature change is that this allows UTF-8 pointers to strings to be
handed to C code again, since the "Latin1 Nowhere" has now been fully
replaced by "UTF-8 Everywhere" implementation.  There are still the
usual caveats about memory relocation (but minus the UNusual caveat of
not knowing how wide a string is...!)