Default left enfix normal to "greedy", add <- operator
The previous rule for left enfix normal finished one complete operation
such that:

    return if 1 > 2 [print "a"] else [print "b"]

...would have ELSE only complete the IF statement before running, not
the return statement.  However, this rule isn't necessarily favorable
in other enfix situations with THEN, OR, AND, etc.

This implements a proposed change to the rules which completes the
entire left by default, but offers an operator to use instead of a
GROUP! with <-:

    return <- if 1 > 2 [print "a"] else [print "b"]