New test format (#766)
This commit changes encoding of individual test vectors to GROUP! and
repurposes BLOCK! to be a containing test group.   Bug references previously
encoded in comments and in ISSUE! are now consistently encoded as an
ISSUE! that begins a test group.  Flags are now encoded as TAG!

The conversion code is temporarily left built-in to the test scripts so that
developers can upgrade their own code bases.  The test parser has been
upgraded to emit token types for precision, documentation and in case further
work is to be done on it.