DOES doesn't intercept RETURN, emulates DO, can specialize
This is a significant change to the semantics and behavior of DOES.
The changes are based on a belief that it is meant to serve a higher
purpose than "being the lazy way to declare a FUNC with no arguments".

Changes to DOES are:

* Unlike FUNC [] [...], DOES [...] has no RETURN
* It soft-quotes its argument
* For types like FILE! / URL! / STRING! it will act as DO when called
* It can be used to specialize functions other than DO
* Blocks you pass to it are LOCKed (if they weren't already)

This contains a somewhat hackish implementation of the specialization
form, that works well enough for most cases.  Ultimately there needs
to be a way to essentially MAKE FRAME! from a VARARGS!...to do one
DO/NEXT worth of evaluation to make a non-stack-based frame.  It's
reaching critical mass that some kind of refactoring will be needed
of that portion of the evaluator logic soon.

At a technical level, DOES of a BLOCK! generates the function
significantly faster than FUNC [], because it does not need to copy it
in most cases.  There are some situations where it will have to do so,
but they are somewhat rare, and it defers the copy until needed.