SELECT/PICK void on miss, ANY/ALL fail on void, DID/NOT voidable
R3-Alpha, Rebol2, and Red conflate these two rather different cases:

    >> select [a b c d] 'e
    == #[none]

    >> select [a b c d e #[none]]
    == #[none]

So an early design rule in Ren-C was to distinguish these, using the
"out of band" nature of void, which was illegal to put in ANY-ARRAY!s.
This enabled a lot of interesting features and (seemingly) helped avoid
bugs in semantics, or at least being alerted to wrong understandings.

But for some casual users, the rarity of "blanks" occurring literally
in their data sets led them to desire a default of "user-friendly"
operations.  The Rebol2 versions of such code would come out seeming
more friendly than the somewhat clunky and laborious versions in Ren-C,
which necessitated handling to keep conditionals from choking on void
variable.  It seemed a step back for simplicity.

Hence a compromise was made: to implement the system primitives in
terms of "accurate" operations (with names like SELECT* and PICK*).
Then "blankifying" operations would be created as CHAINs on these.
Whether someone used SELECT or SELECT*, or PICK or PICK*, etc. would
be based on the application.

However, in the time since that compromise, several advancements were
made in terms of conditional constructs and enfix helpers that could
handle voids.  Written using these tools, the kinds of examples that
were made inelegant previously can generally be made more elegant
than they were before (using ELSE, ALSO, !!, etc.)

To make processing chains of ANY and ALL safer under this dynamic,
this institutes a rule that had been in the air for a while, to error
on void conditions by default.  This is facilitated by having COMMENT,
DUMP, and ELIDE all be able to be "invisible" in a chain of processing,
so most of the reasons one would want a void-returning item to be
allowed in an ANY or ALL are covered by those.  It helps avoid any
accidents that opting out might cause, and restores some base case
behavior for things like ANY [] and ALL [].  New ANY* and ALL*
operations now take the previous behavior of void tolerance.

Also to make things easier, this makes DID and NOT treat void as
logically false.  So people can write `if did x: select ...` and get
the same effect that `if x: select ...` would get before.

Future use of the TRY operation is potentially to serve as a synonym
for TO-VALUE.  That is not done in this commit, but TRY is turned into
a function that directs users to use TRAP instead.