Enable ELIDE, COMMENT, etc. in CASE interstitials
This commit switches CASE to use soft-quoting semantics on its branch
values.  The limiting aspect of this is you can no longer say, e.g.

    3 = case [
       true reduce ['add 1 2]
    ]

If you wanted to say that, you would have to write:

    3 = case [
       true (reduce ['add 1 2])
    ]

There are significant benefits to modifying CASE to this rule.  First,
if a condition is false, it is possible to skip over the branch slot
as it is only one element in size.  This is to say that the following
would not print, and return void:

    void? case [
        false (print "won't print" ['add 1 2])
    ]

While pinning down the semantics of CASE has been a longstanding goal
of Ren-C, the actual motivating scenario was to not "double-evaluate"
branches in the general case, so that ELIDE/DUMP/COMMENT and other
invisibles could be used *between* cases, e.g.:

    flag: false
    result: case [
        1 < 2 [1020]
        elide (flag: true)
        true [fail "shouldn't get here"]
    ]
    (not flag) and (result = 1020)

That is given as an example, because it would have been a problem
in the previous implementation...as flag would be set to TRUE when
the [1020] block was "evaluated".  Now it works as one likely
intends, to only run the elided expression after the previous
condition fails.