Fix --verbose, UNSET?/VALUE? in mezzanine, light cleanup
The ?! operator was written as !? in the --verbose option.  Fixed.

VALUE? had a meaning of looking up a word to see if it was set, while
Ren-C calls that "SET?" and wants to take VALUE? as a synonym of
ANY-VALUE?, e.g. NOT VOID.  It also wants to retake UNSET? as a way
of saying NOT SET?, as VOID? is the way to ask if a cell itself is
not set.

To ease compatibility, these were turned into warnings--so people could
use ANY-VALUE? and NOT SET? until enough time had passed.  Those
warnings are still there, but this goes ahead and makes that something
done in the user context as opposed to the lib context...so that
mezzanine code (bound into lib) can take advantage of it.

This improves some of the boot process, including moving some comments
to Trello cards.  Changes "Ren/C" instances to "Ren-C"