Simplify ASSERT, with VERIFY implemented in usermode
When assert was first written as a usermode hijacking, several features
were added to it...to be able to enable or disable asserts with
ISSUE!s and do other complex things.  This code hadn't been used since
it was written, and was hence not tested.  Rather than carry the
liability of testing and debugging something not mission-critical,
this just coalesces the VERIFY function into usermode along with
the assert.