Make APPLY and MY work with PATH!
This makes APPLY and MY speak the protocol for pushing refinements to
the stack for Do_Core() to process them, by factoring out code that
was introduced for SPECIALIZE.

    >> some-block: copy [a b c]

    >> some-block: my append/dup/part [d e f] 3 2
    == [a b c d e d e d e]

One difficult aspect was that the baseline of the stack needs to be
known, yet Do_Core() may be called on a stack that is constantly
building itself up.  (e.g. REDUCE which pushes a frame and then does
several consecutive evaluations with that frame pushed).  Previously
the DSP was captured on each Do_Core() entry, but this modifies it
so that it gets captured in DECLARE_FRAME(), allowing one to put it
at a separate point from Push_Frame() or a Do_Core() entry.