Address line numbering issues in scanner/headers
R3-Alpha's usermode processing for LOAD would load the file separately
in parts, so the header was obtained through its own TRANSCODE
operation.  However, each transcode assumed line numbering started
over at 0.  Hence error messages would be off by however many lines
were in the header.

The bug was longstanding, and noticed when Ren-C added the FILE-OF and
LINE-OF feature for ANY-ARRAY!.  However, it was deferred to be fixed
at some later time.

This commit attempts to address the issue by slipstreaming the file and
line information of the tail line of the scan onto the BLOCK! which is
the transcode result.  While typically blocks hold the line number of
their open bracket (or groups of their open parenthesis), this block is
only a container--and not used directly by LOAD.

Also corrects a bug in displaying the start line of an unclosed
bracket/parenthesis, and a couple of stray SELECT void issues.