Get rid of R_OUT_UNEVALUATED
When VALUE_FLAG_UNEVALUATED was introduced, there was no Move_Value(),
and cell bits were copied literally.  Now that there is active code
when value cells are moved, that code masks out several bits in the
copy--including NODE_FLAG_MARKED and VALUE_FLAG_UNEVALUATED.  This
means that there is not as much concern over clearing the bit in an
evaluation as before.

Since the bit doesn't need to be cleared, the problem of how a native
like QUOTE can return something with the unevaluated bit goes away.
It doesn't need a special return code to indicate it wants the
evaluator to not clear the bit.