Pass errors back from OS_DO_DEVICE
R3-Alpha had the concept that OS-specific "device code" was supposed
to be agnostic to the point of not knowing about Rebol, and exchanging
all information in C structures vs. using REBVALs themselves through
the API.  This was geared around creating an "abstract OS"--which
could run things other than Rebol.  Ren-C hasn't considered this to
be a relevant idea.

Once the device layer could call Rebol APIs, one of the first ones
that made sense to use would be the error APIs.  This could mean
constructing richer error objects with details about a problem, vs
giving back OS-specific error numbers (or just made up error numbers)
to the device layer.

The first cut of this presumed that the error from the device would
propagate up from the PORT! method directly.  However, that meant the
somewhat raw error messages no longer carried the high level failure
messages that had been historically expected.

This modifies the device call to return the ERROR!, a NULL (for
pending result), or a non-ERROR! result value.  Callsites did not
previously consistently handle DR_PEND, DR_ERROR, and DR_DONE...so
this tries to add in assertions to map the territory.  It is on a
case-by-case basis whether the low level error is retriggered, or if
a higher level error is put in its place.  (An ideal system would
be able to chain the lower level error into the higher one, so that
someone looking for more detail in why an error was caused could
drill down into it.)

Note: Because the device model is largely considered "life support",
most interest is in lifting the evaluator out of it and not using the
code that's there.  However, studying the issues of its interaction
with the core is an example of the kinds of things that would arise
regardless of framework.  These changes focus more on that, vs.
being "pretty" in an objective sense.