Add CONCOCT native, with COMPOSE a specialization of it
This introduces a new twist on COMPOSE speculatively named CONCOCT,
which lets one define a nested pattern of BLOCK! and GROUP! that is
literally searched for in the composition:

    >> concoct (()) [(1 + 2) ((1 + 2))]
    == [(1 + 2) 3]

Originally conceived for making it easier to compose in material to
PARSE rules that already has GROUP!s in it:

    rule: concoct (()) [
        some ["a"] (print "stays as is") ((gets composed))
    ]

However, it can be applied to other cases, such as using BLOCK!s as
the composing element if GROUP!s are more common:

    >> concoct [] quote ((1 + 2) [3 + 4] (5 + 6))
    == ((1 + 2) 7 (5 + 6))

    >> concoct [[]] 'a/(b)/[[1 + 2]]/[c]/d
    == a/(b)/3/[c]/d

This change also makes CONCOCT accept other forms of ANY-ARRAY! as
input, so a PATH! and a GROUP! will be accepted.  /DEEP traversals
will recurse into GROUP!s if the pattern is more complex than a
GROUP!.

Since COMPOSE is a specialization of CONCOCT where pattern is `()`, it
inherits these new behaviors.  Since the type signature of COMPOSE
has changed to not accept non-array types, this should cause errors
that are easy to find.

Regarding not traversing deeply into ANY-PATH! for COMPOSE/DEEP, it
is not clear the original behavior was not an oversight--due to the
relatively late addition of GROUP!s in PATH!s to Rebol2 and R3-Alpha.
The existence of CONCOCT makes it easier to work around this if there
are GROUP!s that truly aren't meant to be touched.

Note: When originally concieved, the patterns could be even more
interesting, such as:

    >> concoct (**) [(1 + 2) (** 3 + 4 **)]
    == [(1 + 2) 7]

The idea being that the symmetric patterns would be omitted from the
evaluation.  This would be an interesting idea, except that there is
no DO/PART feature...the evaluator can be told to start at a certain
place in an array and not look back, but it can't be blinded past
a certain point.  So in the above example, the terminal ** would
interfere in the evaluation.