Don't rebind spliced REBVALs in API, insulate console context
As a quick demo to show the API running spliced runs of text and
REBVAL*s together from C, some answer to binding the text runs had
to be done.  The quick and dirty solution was that when a C char*
string was seen in a va_list (recognized via legal UTF-8 leading byte)
then the ensuing scanned values would be bound the way INTERN does...
to user first, and then to interned versions of things from lib
secondly.

This known bad strategy could not distinguish the spliced values (which
may have had their own bindings, meant to be preserved) from data
loaded from the raw character segments.  Hence an operation like
`rebDid("lib/error?", code)` would wipe out any bindings on code...
which may be relevant (e.g. in the console).

As a result of the loss of these bindings, console code was being
rebound into user space...losing its protections.  This was masked
because the console itself was loaded into `lib` and hence was
interned like any lib function would be.  So instead of suddenly
seeing bindings go missing, they just appeared to keep working, but
could be overwritten easily.

This commit addresses both problems in a way that isn't perfect, but
certainly much better.  The console is no longer loaded into lib, but
gets its own context.  And the scanner is now parameterized with a
binder so it can intern code as it is loaded...which is a useful
feature that can likely speed up other parts of the system.

It also corrects a bug in HALT, and a GC problem in printing where a
temporary UTF-8 series was not being guarded during time when it
could be in use.